Module Tezos_010_PtGRANAD_test_helpers.Account

type t = {

pkh : Tezos_crypto.Signature.Public_key_hash.t;

pk : Tezos_crypto.Signature.Public_key.t;

sk : Tezos_crypto.Signature.Secret_key.t;

}

type account = t

val known_accounts : t Tezos_crypto.Signature.Public_key_hash.Table.t

val activator_account : account

val dummy_account : account

val new_account : ?seed:Stdlib.Bytes.t -> unit -> account

val add_account : t -> unit

val find : Tezos_crypto.Signature.Public_key_hash.t -> (t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val find_alternate : Tezos_crypto.Signature.Public_key_hash.t -> t

generate_accounts ?initial_balances n : generates n random accounts with the initial balance of the ith account given by the ith value in the list initial_balances or otherwise 4.000.000.000 tz (if the list is too short); and add them to the global account state

val generate_accounts : ?rng_state:Stdlib.Random.State.t -> ?initial_balances:int64 list -> int -> (t * Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Tez.t) list

val commitment_secret : Tezos_protocol_010_PtGRANAD.Protocol.Blinded_public_key_hash.activation_code

val new_commitment : ?seed:Stdlib.Bytes.t -> unit -> (account * Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Commitment.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t