Module Context.Contract

val pp : Tezos_protocol_010_PtGRANAD.Protocol.Environment.Format.formatter -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> unit

val equal : Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> bool

val pkh : Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

type balance_kind =

| Main

| Deposit

| Fees

| Rewards

val balance : ?⁠kind:balance_kind -> t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Tez.t Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

Returns the balance of a contract, by default the main balance. If the contract is implicit the frozen balances are available too: deposit, fees or rewards.

val counter : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_protocol_010_PtGRANAD.Protocol.Environment.Z.t Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val manager : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Account.t Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val is_manager_key_revealed : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> bool Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val delegate : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val delegate_opt : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash option Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val storage : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Script.expr Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val script : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Script.expr Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t

val script_hash : t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Script_expr_hash.t Tezos_base__TzPervasives.tzresult Tezos_protocol_010_PtGRANAD.Protocol.Environment.Lwt.t