Module Sapling_helpers.Interpreter_helpers

include Common

val memo_size_of_int : int -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Sapling.Memo_size.t

val int_of_memo_size : Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Sapling.Memo_size.t -> int

val wrap : 'a Tezos_protocol_010_PtGRANAD.Protocol.Environment.Error_monad.tzresult -> 'a Tezos_base__TzPervasives.tzresult Lwt.t

val assert_true : (bool, 'a) Stdlib.result Lwt.t -> (unit, 'a) Stdlib.result Lwt.t

val assert_false : (bool, 'a) Stdlib.result Lwt.t -> (unit, 'a) Stdlib.result Lwt.t

val assert_some : ('a option, 'b) Stdlib.result Lwt.t -> ('a, 'b) Stdlib.result Lwt.t

val assert_none : ('a option, 'b) Stdlib.result Lwt.t -> (unit, 'b) Stdlib.result Lwt.t

val assert_error : ('a, 'b) Stdlib.result Lwt.t -> (unit, 'c) Stdlib.result Lwt.t

val print : ?⁠prefix:string -> 'a Data_encoding.Encoding.t -> 'a -> unit

val to_hex : 'a -> 'a Data_encoding.Encoding.t -> string

val randomized_byte : ?⁠pos:int -> 'a -> 'a Data_encoding.Encoding.t -> 'a

type wallet = {

sk : Tezos_sapling.Core.Wallet.Spending_key.t;

vk : Tezos_sapling.Core.Wallet.Viewing_key.t;

}

val wallet_gen : unit -> wallet

val gen_addr : int -> Tezos_sapling.Core.Client.Viewing_key.t -> Tezos_sapling.Core.Client.Viewing_key.address list

val gen_nf : unit -> Tezos_sapling.Core.Client.Nullifier.t

val gen_cm_cipher : memo_size:int -> unit -> Tezos_sapling.Core.Client.Commitment.t * Tezos_sapling.Core.Client.Ciphertext.t

val client_state_of_diff : memo_size:int -> (Tezos_sapling.Core.Validator.Hash.t * Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Sapling.diff) -> Tezos_sapling.Storage.state

include Contract_helpers

val init : unit -> (Block.t * Tezos_raw_protocol_010_PtGRANAD__Alpha_context.public_key_hash * Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t * Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

Initializes 2 addresses to do only operations plus one that will be used to bake.

val toplevel_from_string : string -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Script.expr

Parses a Michelson contract from string.

val expression_from_string : string -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Script.expr

Parses a Michelson expression from string, useful for call parameters.

val originate_contract : string -> string -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.contract -> Block.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash -> (Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.contract * Block.t, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

Returns a block in which the contract is originated.

val originate_contract : string -> string -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.contract -> Block.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash -> (Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.contract * Block.t * string, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

Returns a block in which the contract is originated. Also returns the associated anti-replay string and KT1 address.

val hex_shield : memo_size:int -> wallet -> string -> string

val transac_and_sync : memo_size:int -> Block.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Script.lazy_expr -> int -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Contract.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.public_key_hash -> (Block.t * Tezos_raw_protocol_010_PtGRANAD.Alpha_context.context * Tezos_sapling.Storage.state, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t

val shield : memo_size:int -> Tezos_sapling.Forge.Core.Spending_key.t -> int -> Tezos_sapling.Core.Wallet.Viewing_key.t -> (string -> 'a) -> string -> 'a list * int

val next_block : Block.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Operation.packed -> (Block.t, Tezos_base__TzPervasives.tztrace) Stdlib.result Lwt.t