Module Tezos_010_PtGRANAD_test_helpers.Contract_helpers

val init : unit -> (Block.t * Tezos_crypto.Signature.public_key_hash * Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t * Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Initializes 2 addresses to do only operations plus one that will be used to bake.

val toplevel_from_string : string -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Script.expr

Parses a Michelson contract from string.

val expression_from_string : string -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Script.expr

Parses a Michelson expression from string, useful for call parameters.

val originate_contract : string -> string -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.contract -> Block.t -> Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.public_key_hash -> (Tezos_protocol_010_PtGRANAD.Protocol.Alpha_context.Contract.contract * Block.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

Returns a block in which the contract is originated.